+++ src/blocks/pullquote/pullquote.js
@@ -14,6 +14,7 @@ import './editor.scss';
 // Internal dependencies.
 import Background, { BackgroundAttributes } from '../../components/background';
 import getAllowedFormats from '../../global/allowed-formats';
+import themeOptions from '../../global/theme-options';
 
 // WordPress dependencies.
 const {
@@ -33,7 +34,9 @@ const {
 } = wp.components;
 const {
 	InspectorControls,
+	PanelColorSettings,
 	RichText,
+	withColors,
 } = ( 'undefined' === typeof wp.blockEditor ) ? wp.editor : wp.blockEditor;
 
 // Returns true if the current block style is "Default".
@@ -41,8 +44,15 @@ const isStyleDefault = ( className ) => {
 	return ( ! className.includes( 'is-style-modern' ) && ! className.includes( 'is-style-pop' ) );
 };
 
-// Returns the class list for the block based on the current settings.
-const getClasses = ( className, backgroundId, imageFocus ) => {
+/**
+ * Returns the class list for the block based on the current settings.
+ *
+ * @param {string} className    The classnames assigned to the block
+ * @param {number} backgroundId ID of the background media, if set.
+ * @param {string} imageFocus   Value of the "Crop Media To" setting.
+ * @param {string} themeColor   Value of the "Theme Color" setting.
+ */
+const getClasses = ( className, backgroundId, imageFocus, themeColor ) => {
 	const isStylePop = className.includes( 'is-style-pop' );
 
 	return (
@@ -51,6 +61,7 @@ const getClasses = ( className, backgroundId, imageFocus ) => {
 			{
 				'has-image': ( backgroundId && ! isStylePop ),
 				[ `has-image-focus-${imageFocus}` ]: ( imageFocus && ! isStylePop ),
+				[ `has-${themeColor}-theme` ]: themeColor,
 			}
 		)
 	);
@@ -86,6 +97,10 @@ registerBlockType( 'bu/pullquote', {
 		className: {
 			type: 'string',
 		},
+		themeColor: {
+			type: 'string',
+			default: '',
+		},
 		...BackgroundAttributes,
 	},
 	styles: [
@@ -104,12 +119,14 @@ registerBlockType( 'bu/pullquote', {
 		},
 	],
 
-	edit( props ) {
+	edit: withColors( 'themeColor' )( props => {
 		// Get the block properties.
 		const {
 			attributes,
 			setAttributes,
 			className,
+			setThemeColor,
+			themeColor,
 		} = props;
 
 		// Get the block attributes.
@@ -153,9 +170,22 @@ registerBlockType( 'bu/pullquote', {
 		return (
 			<Fragment>
 				<InspectorControls>
+					<PanelColorSettings
+						title={ __( 'Theme Color' ) }
+						initialOpen={ false }
+						colorSettings={ [
+							{
+								value: themeColor.color,
+								onChange: setThemeColor,
+								label: __( 'Theme' ),
+								disableCustomColors: true,
+								colors: themeOptions(),
+							},
+						] }
+					/>
 					{ mediaPositioningControls() }
 				</InspectorControls>
-				<div className={ getClasses( className, backgroundId, imageFocus ) }>
+				<div className={ getClasses( className, backgroundId, imageFocus, themeColor.slug ) }>
 					{ isStyleDefault( className ) && (
 						<Background
 							allowedMediaTypes={ allowedMedia }
@@ -204,7 +234,7 @@ registerBlockType( 'bu/pullquote', {
 				</div>
 			</Fragment>
 		);
-	},
+	} ),
 
 	save( props ) {
 		// Get the block properties.
@@ -219,11 +249,12 @@ registerBlockType( 'bu/pullquote', {
 			imageFocus,
 			backgroundId,
 			className = '', // Assign default in case the unpacked value is `undefined`.
+			themeColor,
 		} = attributes;
 
 		// Returns the block rendering for the front end.
 		return (
-			<div className={ getClasses( className, backgroundId, imageFocus ) }>
+			<div className={ getClasses( className, backgroundId, imageFocus, themeColor ) }>
 				{ isStyleDefault( className ) && (
 					<figure>
 						<Background